cleans up types
